wip: further refactoring into functional style module
